adding gitpull command

+++ main.py
@@ -105,8 +105,8 @@ async def gitpull(ctx: commands.Context):
     message = await ctx.send('Pulling latest files from repo...')
     repo = git.Repo('/home/ubuntu/Snowphia/.git')
     ret = repo.remotes.origin.pull()
-    await ctx.send(str(ret))
-    await ctx.send(str(ret[0].flags))
+    if ret[0].flags == 4:
+        return
     await message.delete()
     command = client.get_command('restart')
     await ctx.invoke(command)
